chore: simplified docker setup
Now docker image is built on alpine (lighter) and supports stages. It also includes composer and xdebug out of the box. This mean the whole development environment is contained there.

Also, for testing purposes we don't need to run apache. The built in web server in PHP is more than enough.